+++ Xplat/src/main/java/vazkii/botania/common/item/CorporeaSparkItem.java
@@ -12,6 +12,7 @@ import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.item.DyeItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
@@ -21,6 +22,7 @@ import net.minecraft.world.level.Level;
 import vazkii.botania.api.corporea.CorporeaHelper;
 import vazkii.botania.common.entity.BotaniaEntities;
 import vazkii.botania.common.entity.CorporeaSparkEntity;
+import vazkii.botania.common.helper.EntityHelper;
 import vazkii.botania.common.impl.corporea.DummyCorporeaNode;
 import vazkii.botania.common.lib.BotaniaTags;
 
@@ -28,14 +30,21 @@ import java.util.List;
 
 public class CorporeaSparkItem extends Item {
 
-	public CorporeaSparkItem(Properties props) {
-		super(props);
+	public CorporeaSparkItem(Properties properties) {
+		super(properties);
 	}
 
 	@Override
-	public InteractionResult useOn(UseOnContext ctx) {
-		return attachSpark(ctx.getLevel(), ctx.getClickedPos(), ctx.getItemInHand())
-				? InteractionResult.sidedSuccess(ctx.getLevel().isClientSide())
+	public InteractionResult useOn(UseOnContext context) {
+		ItemStack otherHandStack = ItemStack.EMPTY;
+		if (context.getPlayer() != null) {
+			otherHandStack = context.getPlayer().getItemInHand(EntityHelper.otherHand(context.getHand()));
+			if (context.getPlayer().isCreative()) {
+				otherHandStack = otherHandStack.copy();
+			}
+		}
+		return attachSpark(context.getLevel(), context.getClickedPos(), context.getItemInHand(), otherHandStack)
+				? InteractionResult.sidedSuccess(context.getLevel().isClientSide())
 				: InteractionResult.PASS;
 	}
 
@@ -44,7 +53,7 @@ public class CorporeaSparkItem extends Item {
 				|| !(spark.getSparkNode() instanceof DummyCorporeaNode);
 	}
 
-	public static boolean attachSpark(Level world, BlockPos pos, ItemStack stack) {
+	public static boolean attachSpark(Level world, BlockPos pos, ItemStack stack, ItemStack otherHandStack) {
 		CorporeaSparkEntity spark = BotaniaEntities.CORPOREA_SPARK.create(world);
 		if (stack.is(BotaniaItems.MASTER_CORPOREA_SPARK)) {
 			spark.setMaster(true);
@@ -53,6 +62,10 @@ public class CorporeaSparkItem extends Item {
 			spark.setCreative(true);
 		}
 		spark.setPos(pos.getX() + 0.5, pos.getY() + 1.25, pos.getZ() + 0.5);
+		if (otherHandStack.getItem() instanceof DyeItem dye) {
+			otherHandStack.shrink(1);
+			spark.setNetwork(dye.getDyeColor());
+		}
 
 		if (canPlace(world, spark) && !CorporeaHelper.instance().doesBlockHaveSpark(world, pos)) {
 			if (!world.isClientSide) {