+++ Xplat/src/main/java/vazkii/botania/common/item/ManaSparkItem.java
@@ -10,6 +10,7 @@ package vazkii.botania.common.item;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.item.DyeItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.UseOnContext;
@@ -18,21 +19,29 @@ import net.minecraft.world.level.Level;
 import vazkii.botania.api.mana.spark.ManaSparkAttachable;
 import vazkii.botania.api.mana.spark.ManaSparkHelper;
 import vazkii.botania.common.entity.ManaSparkEntity;
+import vazkii.botania.common.helper.EntityHelper;
 
 public class ManaSparkItem extends Item {
 
-	public ManaSparkItem(Properties builder) {
-		super(builder);
+	public ManaSparkItem(Properties properties) {
+		super(properties);
 	}
 
 	@Override
-	public InteractionResult useOn(UseOnContext ctx) {
-		return attachSpark(ctx.getLevel(), ctx.getClickedPos(), ctx.getItemInHand())
-				? InteractionResult.sidedSuccess(ctx.getLevel().isClientSide())
+	public InteractionResult useOn(UseOnContext context) {
+		ItemStack otherHandStack = ItemStack.EMPTY;
+		if (context.getPlayer() != null) {
+			otherHandStack = context.getPlayer().getItemInHand(EntityHelper.otherHand(context.getHand()));
+			if (context.getPlayer().isCreative()) {
+				otherHandStack = otherHandStack.copy();
+			}
+		}
+		return attachSpark(context.getLevel(), context.getClickedPos(), context.getItemInHand(), otherHandStack)
+				? InteractionResult.sidedSuccess(context.getLevel().isClientSide())
 				: InteractionResult.PASS;
 	}
 
-	public static boolean attachSpark(Level world, BlockPos pos, ItemStack stack) {
+	public static boolean attachSpark(Level world, BlockPos pos, ItemStack stack, ItemStack otherHandStack) {
 		var attach = ManaSparkAttachable.LOOKUP.find(world, pos);
 		if (attach != null) {
 			if (attach.canAttachSpark(stack) && ManaSparkHelper.getAttachedSpark(world, pos) == null) {
@@ -40,6 +49,10 @@ public class ManaSparkItem extends Item {
 					stack.shrink(1);
 					ManaSparkEntity spark = new ManaSparkEntity(world);
 					spark.setPos(pos.getX() + 0.5, pos.getY() + 1.25, pos.getZ() + 0.5);
+					if (otherHandStack.getItem() instanceof DyeItem dye) {
+						otherHandStack.shrink(1);
+						spark.setNetwork(dye.getDyeColor());
+					}
 					world.addFreshEntity(spark);
 					attach.attachSpark(spark);
 				}