+++ web/changelog.md
@@ -86,6 +86,7 @@ Here's a general overview of what you can expect to change in terms of gameplay:
 The following changes will be released in an upcoming version of Botania for 1.20.1:
 
 * Add: Manaseer Monocle on the Vinculotus shows where Endermen will teleport to
+* Add: Placing sparks while holding dye in your other hand will dye the spark that color
 * Change: Hopperhock no longer shows a secondary range indicator, as it doesn't really match its container access pattern
 * Change: Ender Overseer line of sight check type switched from collision to visual (like 3rd-person camera or trial spawners)
 * Remove: Quark frog as rare cocoon spawn (catching up to 1.19 – woo, so up-to-date!)